+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/laf/package-info.java
@@ -18,6 +18,8 @@
  */
 /**
  * Package configuration values:<br>
+ * {@link is.codion.swing.common.ui.laf.LookAndFeelProvider#CROSS_PLATFORM}
+ * {@link is.codion.swing.common.ui.laf.LookAndFeelProvider#SYSTEM}
  * {@link is.codion.swing.common.ui.laf.LookAndFeelComboBox#ENABLE_ON_SELECTION}
  */
 package is.codion.swing.common.ui.laf;